change test framework to mocha, change typescript compiler to babel
