+++ examples/__tests__/deviantart.unit.test.ts
@@ -2,7 +2,7 @@ import deviantartConfig from '../deviantart.config.json'
 import { assertConfigType } from '../../src/configuration/site-traversal'
 
 describe('deviantart config', () => {
-  test('is properly typed', () => {
+  it('is properly typed', () => {
     assertConfigType(deviantartConfig)
   })
