+++ examples/__tests__/nasa-image-of-the-day.unit.test.ts
@@ -2,10 +2,9 @@ import nasaIotdConfig from '../nasa-image-of-the-day.config.json'
 import { assertConfigType } from '../../src/configuration/site-traversal'
 
 describe('nasa iotd config', () => {
-  test('is properly typed', () => {
+  it('is properly typed', () => {
     assertConfigType(nasaIotdConfig)
   })
 
-  // TODO
-  // test('returns expected example output', () => {})
+  it('returns expected example output', () => {})
 })
