+++ examples/__tests__/tumblr.unit.test.ts
@@ -2,7 +2,7 @@ import tumblrConfig from '../tumblr.config.json'
 import { assertConfigType } from '../../src/configuration/site-traversal'
 
 describe('tumblr config', () => {
-  test('is properly typed', () => {
+  it('is properly typed', () => {
     assertConfigType(tumblrConfig)
   })
 })
