+++ package.json
@@ -12,14 +12,14 @@
     "build": "run-s clean compile:runtime compile",
     "build:watch": "webpack --mode=development --watch",
     "compile": "webpack --mode=production",
-    "compile:runtime": "tsr --compilerOptions '{\"outDir\": \"src/configuration/runtime/\"}' src/configuration/site-traversal/assert.ts && tsr --compilerOptions '{\"outDir\": \"src/run-options/runtime\"}' src/configuration/run-options/assert.ts",
+    "compile:runtime": "tsr --compilerOptions '{\"outDir\": \"src/configuration/site-traversal/runtime/\"}' src/configuration/site-traversal/assert.ts && tsr --compilerOptions '{\"outDir\": \"src/configuration/run-options/runtime\"}' src/configuration/run-options/assert.ts",
     "clean": "rimraf lib",
     "lint": "git ls-files '*.ts' '*.json' | xargs prettier -l",
     "typecheck": "tsc --noEmit --project tsconfig.json",
     "typecheck:watch": "npm run typecheck -- --watch",
-    "test": "jest",
-    "test:unit": "jest unit",
-    "test:integrations": "jest integration --detectOpenHandles"
+    "test": "mocha-webpack",
+    "test:unit": "npm run test **/*.unit.test.ts",
+    "test:integrations": "npm run test integration --detectOpenHandles"
   },
   "repository": {
     "type": "git",
@@ -31,36 +31,33 @@
     "node": "<=8.0"
   },
   "devDependencies": {
+    "@babel/core": "^7.2.2",
+    "@babel/plugin-proposal-class-properties": "^7.2.3",
+    "@babel/preset-env": "^7.2.3",
+    "@babel/preset-typescript": "^7.1.0",
+    "@babel/register": "^7.0.0",
     "@types/better-sqlite3": "^3.1.3",
+    "@types/chai": "^4.1.7",
     "@types/cheerio": "^0.22.9",
     "@types/handlebars": "^4.0.39",
-    "@types/jest": "^23.3.3",
+    "@types/mocha": "^5.2.5",
+    "@types/nock": "^9.3.0",
     "@types/node-fetch": "^2.1.2",
     "@types/sanitize-filename": "^1.1.28",
     "@types/source-map-support": "^0.4.1",
-    "babel-core": "^6.26.3",
-    "babel-jest": "^23.0.1",
-    "babel-loader": "^7.1.4",
-    "babel-plugin-add-module-exports": "^0.3.2",
-    "babel-plugin-flow-runtime": "^0.17.0",
-    "babel-plugin-transform-flow-strip-types": "^6.22.0",
-    "babel-preset-env": "^1.7.0",
-    "babel-preset-stage-2": "^6.24.1",
+    "babel-loader": "^8.0.4",
+    "chai": "^4.2.0",
     "clean-terminal-webpack-plugin": "^1.1.0",
     "copy-webpack-plugin": "^4.5.2",
-    "flow-bin": "^0.78.0",
-    "fork-ts-checker-webpack-plugin": "^0.4.9",
-    "jest": "^23.1.0",
-    "jest-raw-loader": "^1.0.1",
+    "mocha": "^5.2.0",
+    "mocha-webpack": "^2.0.0-beta.0",
+    "nock": "^10.0.3",
     "npm-run-all": "^4.1.5",
     "prettier": "^1.13.5",
     "raw-loader": "^0.5.1",
     "rimraf": "^2.6.2",
     "rxjs-marbles": "^4.3.5",
     "terser-webpack-plugin": "^1.0.0",
-    "ts-jest": "^23.10.3",
-    "ts-loader": "^5.2.1",
-    "ts-runtime": "^0.1.35",
     "typescript": "^3.1.1",
     "webpack": "^4.12.0",
     "webpack-cli": "^3.0.6",
@@ -69,36 +66,11 @@
   "dependencies": {
     "better-sqlite3": "^4.1.4",
     "cheerio": "^1.0.0-rc.2",
-    "flow-runtime": "^0.17.0",
     "handlebars": "^4.0.12",
     "node-fetch": "^2.1.2",
     "rxjs": "^6.3.3",
     "sanitize-filename": "^1.6.1",
-    "source-map-support": "^0.5.6"
-  },
-  "jest": {
-    "testEnvironment": "node",
-    "setupFiles": [
-      "./tests/setup.ts"
-    ],
-    "verbose": true,
-    "moduleDirectories": [
-      "node_modules",
-      "src"
-    ],
-    "roots": [
-      "<rootDir>/examples",
-      "<rootDir>/src"
-    ],
-    "transform": {
-      "\\.sql$": "jest-raw-loader",
-      "\\.(ts|js|json)$": "ts-jest"
-    },
-    "testRegex": "(/__tests__/.*|(\\.|/)(test))\\.ts$",
-    "moduleFileExtensions": [
-      "ts",
-      "js",
-      "json"
-    ]
+    "source-map-support": "^0.5.6",
+    "ts-runtime": "^0.2.0"
   }
 }
