--- src/configuration/runtime/tsr-declarations.js
@@ -1,20 +1,0 @@
-"use strict";
-var _this = this;
-Object.defineProperty(exports, "__esModule", { value: true });
-var lib_1 = require("ts-runtime/lib");
-lib_1.default.declare("Array.3170645078", lib_1.default.type("Array", function (Array) {
-    var T = Array.typeParameter("T");
-    return lib_1.default.object(lib_1.default.property("length", lib_1.default.nullable(lib_1.default.number())), lib_1.default.indexer("n", lib_1.default.nullable(lib_1.default.number()), lib_1.default.nullable(T)), lib_1.default.property("toString", lib_1.default.nullable(lib_1.default.function(lib_1.default.return(lib_1.default.nullable(lib_1.default.string()))))), lib_1.default.property("toLocaleString", lib_1.default.nullable(lib_1.default.function(lib_1.default.return(lib_1.default.nullable(lib_1.default.string()))))), lib_1.default.property("push", lib_1.default.nullable(lib_1.default.function(lib_1.default.param("items", lib_1.default.nullable(lib_1.default.array(lib_1.default.nullable(T))), true), lib_1.default.return(lib_1.default.nullable(lib_1.default.number()))))), lib_1.default.property("pop", lib_1.default.nullable(lib_1.default.function(lib_1.default.return(lib_1.default.nullable(lib_1.default.union(lib_1.default.nullable(T), lib_1.default.nullable(lib_1.default.undef()))))))), lib_1.default.property("concat", lib_1.default.nullable(lib_1.default.function(lib_1.default.param("items", lib_1.default.nullable(lib_1.default.union(lib_1.default.nullable(lib_1.default.array(lib_1.default.nullable(lib_1.default.ref(ConcatArray, lib_1.default.nullable(T))))), lib_1.default.nullable(lib_1.default.array(lib_1.default.nullable(lib_1.default.union(lib_1.default.nullable(T), lib_1.default.nullable(lib_1.default.ref(ConcatArray, lib_1.default.nullable(T))))))))), true), lib_1.default.return(lib_1.default.nullable(lib_1.default.array(lib_1.default.nullable(T))))))), lib_1.default.property("join", lib_1.default.nullable(lib_1.default.function(lib_1.default.param("separator", lib_1.default.nullable(lib_1.default.string()), true), lib_1.default.return(lib_1.default.nullable(lib_1.default.string()))))), lib_1.default.property("reverse", lib_1.default.nullable(lib_1.default.function(lib_1.default.return(lib_1.default.nullable(lib_1.default.array(lib_1.default.nullable(T))))))), lib_1.default.property("shift", lib_1.default.nullable(lib_1.default.function(lib_1.default.return(lib_1.default.nullable(lib_1.default.union(lib_1.default.nullable(T), lib_1.default.nullable(lib_1.default.undef()))))))), lib_1.default.property("slice", lib_1.default.nullable(lib_1.default.function(lib_1.default.param("start", lib_1.default.nullable(lib_1.default.number()), true), lib_1.default.param("end", lib_1.default.nullable(lib_1.default.number()), true), lib_1.default.return(lib_1.default.nullable(lib_1.default.array(lib_1.default.nullable(T))))))), lib_1.default.property("sort", lib_1.default.nullable(lib_1.default.function(lib_1.default.param("compareFn", lib_1.default.nullable(lib_1.default.function(lib_1.default.param("a", lib_1.default.nullable(T)), lib_1.default.param("b", lib_1.default.nullable(T)), lib_1.default.return(lib_1.default.nullable(lib_1.default.number())))), true), lib_1.default.return(lib_1.default.nullable(lib_1.default.this(_this)))))), lib_1.default.property("splice", lib_1.default.nullable(lib_1.default.function(lib_1.default.param("start", lib_1.default.nullable(lib_1.default.number())), lib_1.default.param("deleteCount", lib_1.default.nullable(lib_1.default.number()), true), lib_1.default.param("items", lib_1.default.nullable(lib_1.default.array(lib_1.default.nullable(T))), true), lib_1.default.return(lib_1.default.nullable(lib_1.default.array(lib_1.default.nullable(T))))))), lib_1.default.property("unshift", lib_1.default.nullable(lib_1.default.function(lib_1.default.param("items", lib_1.default.nullable(lib_1.default.array(lib_1.default.nullable(T))), true), lib_1.default.return(lib_1.default.nullable(lib_1.default.number()))))), lib_1.default.property("indexOf", lib_1.default.nullable(lib_1.default.function(lib_1.default.param("searchElement", lib_1.default.nullable(T)), lib_1.default.param("fromIndex", lib_1.default.nullable(lib_1.default.number()), true), lib_1.default.return(lib_1.default.nullable(lib_1.default.number()))))), lib_1.default.property("lastIndexOf", lib_1.default.nullable(lib_1.default.function(lib_1.default.param("searchElement", lib_1.default.nullable(T)), lib_1.default.param("fromIndex", lib_1.default.nullable(lib_1.default.number()), true), lib_1.default.return(lib_1.default.nullable(lib_1.default.number()))))), lib_1.default.property("every", lib_1.default.nullable(lib_1.default.function(lib_1.default.param("callbackfn", lib_1.default.nullable(lib_1.default.function(lib_1.default.param("value", lib_1.default.nullable(T)), lib_1.default.param("index", lib_1.default.nullable(lib_1.default.number())), lib_1.default.param("array", lib_1.default.nullable(lib_1.default.array(lib_1.default.nullable(T)))), lib_1.default.return(lib_1.default.nullable(lib_1.default.boolean()))))), lib_1.default.param("thisArg", lib_1.default.any(), true), lib_1.default.return(lib_1.default.nullable(lib_1.default.boolean()))))), lib_1.default.property("some", lib_1.default.nullable(lib_1.default.function(lib_1.default.param("callbackfn", lib_1.default.nullable(lib_1.default.function(lib_1.default.param("value", lib_1.default.nullable(T)), lib_1.default.param("index", lib_1.default.nullable(lib_1.default.number())), lib_1.default.param("array", lib_1.default.nullable(lib_1.default.array(lib_1.default.nullable(T)))), lib_1.default.return(lib_1.default.nullable(lib_1.default.boolean()))))), lib_1.default.param("thisArg", lib_1.default.any(), true), lib_1.default.return(lib_1.default.nullable(lib_1.default.boolean()))))), lib_1.default.property("forEach", lib_1.default.nullable(lib_1.default.function(lib_1.default.param("callbackfn", lib_1.default.nullable(lib_1.default.function(lib_1.default.param("value", lib_1.default.nullable(T)), lib_1.default.param("index", lib_1.default.nullable(lib_1.default.number())), lib_1.default.param("array", lib_1.default.nullable(lib_1.default.array(lib_1.default.nullable(T)))), lib_1.default.return(lib_1.default.void())))), lib_1.default.param("thisArg", lib_1.default.any(), true), lib_1.default.return(lib_1.default.void())))), lib_1.default.property("map", lib_1.default.nullable(lib_1.default.function(function (fn) {
-        var U = fn.typeParameter("U");
-        return [lib_1.default.param("callbackfn", lib_1.default.nullable(lib_1.default.function(lib_1.default.param("value", lib_1.default.nullable(T)), lib_1.default.param("index", lib_1.default.nullable(lib_1.default.number())), lib_1.default.param("array", lib_1.default.nullable(lib_1.default.array(lib_1.default.nullable(T)))), lib_1.default.return(lib_1.default.nullable(U))))), lib_1.default.param("thisArg", lib_1.default.any(), true), lib_1.default.return(lib_1.default.nullable(lib_1.default.array(lib_1.default.nullable(U))))];
-    }))), lib_1.default.property("filter", lib_1.default.nullable(lib_1.default.function(function (fn) {
-        var S = fn.typeParameter("S", lib_1.default.nullable(T));
-        return [lib_1.default.param("callbackfn", lib_1.default.nullable(lib_1.default.union(lib_1.default.nullable(lib_1.default.function(lib_1.default.param("value", lib_1.default.nullable(T)), lib_1.default.param("index", lib_1.default.nullable(lib_1.default.number())), lib_1.default.param("array", lib_1.default.nullable(lib_1.default.array(lib_1.default.nullable(T)))), lib_1.default.return(lib_1.default.nullable(lib_1.default.boolean())))), lib_1.default.nullable(lib_1.default.function(lib_1.default.param("value", lib_1.default.nullable(T)), lib_1.default.param("index", lib_1.default.nullable(lib_1.default.number())), lib_1.default.param("array", lib_1.default.nullable(lib_1.default.array(lib_1.default.nullable(T)))), lib_1.default.return(lib_1.default.any())))))), lib_1.default.param("thisArg", lib_1.default.any(), true), lib_1.default.return(lib_1.default.nullable(lib_1.default.union(lib_1.default.nullable(lib_1.default.array(lib_1.default.nullable(S))), lib_1.default.nullable(lib_1.default.array(lib_1.default.nullable(T))))))];
-    }))), lib_1.default.property("reduce", lib_1.default.nullable(lib_1.default.function(function (fn) {
-        var U = fn.typeParameter("U");
-        return [lib_1.default.param("callbackfn", lib_1.default.nullable(lib_1.default.union(lib_1.default.nullable(lib_1.default.function(lib_1.default.param("previousValue", lib_1.default.nullable(T)), lib_1.default.param("currentValue", lib_1.default.nullable(T)), lib_1.default.param("currentIndex", lib_1.default.nullable(lib_1.default.number())), lib_1.default.param("array", lib_1.default.nullable(lib_1.default.array(lib_1.default.nullable(T)))), lib_1.default.return(lib_1.default.nullable(T)))), lib_1.default.nullable(lib_1.default.function(lib_1.default.param("previousValue", lib_1.default.nullable(U)), lib_1.default.param("currentValue", lib_1.default.nullable(T)), lib_1.default.param("currentIndex", lib_1.default.nullable(lib_1.default.number())), lib_1.default.param("array", lib_1.default.nullable(lib_1.default.array(lib_1.default.nullable(T)))), lib_1.default.return(lib_1.default.nullable(U))))))), lib_1.default.param("initialValue", lib_1.default.nullable(lib_1.default.union(lib_1.default.nullable(T), lib_1.default.nullable(U))), true), lib_1.default.return(lib_1.default.nullable(lib_1.default.union(lib_1.default.nullable(T), lib_1.default.nullable(U))))];
-    }))), lib_1.default.property("reduceRight", lib_1.default.nullable(lib_1.default.function(function (fn) {
-        var U = fn.typeParameter("U");
-        return [lib_1.default.param("callbackfn", lib_1.default.nullable(lib_1.default.union(lib_1.default.nullable(lib_1.default.function(lib_1.default.param("previousValue", lib_1.default.nullable(T)), lib_1.default.param("currentValue", lib_1.default.nullable(T)), lib_1.default.param("currentIndex", lib_1.default.nullable(lib_1.default.number())), lib_1.default.param("array", lib_1.default.nullable(lib_1.default.array(lib_1.default.nullable(T)))), lib_1.default.return(lib_1.default.nullable(T)))), lib_1.default.nullable(lib_1.default.function(lib_1.default.param("previousValue", lib_1.default.nullable(U)), lib_1.default.param("currentValue", lib_1.default.nullable(T)), lib_1.default.param("currentIndex", lib_1.default.nullable(lib_1.default.number())), lib_1.default.param("array", lib_1.default.nullable(lib_1.default.array(lib_1.default.nullable(T)))), lib_1.default.return(lib_1.default.nullable(U))))))), lib_1.default.param("initialValue", lib_1.default.nullable(lib_1.default.union(lib_1.default.nullable(T), lib_1.default.nullable(U))), true), lib_1.default.return(lib_1.default.nullable(lib_1.default.union(lib_1.default.nullable(T), lib_1.default.nullable(U))))];
-    }))));
-}));
