+++ src/index.ts
@@ -1,5 +1,4 @@
+import 'source-map-support/register'
 import PageScraper from './scraper'
-import sourceMapSupport from 'source-map-support'
-sourceMapSupport.install()
 
 export default PageScraper
