+++ src/scraper/__tests__/index.unit.test.ts
@@ -4,7 +4,7 @@ import * as globalVals from '../../../tests/setup'
 describe('scrape-pages class', () => {
   const simpleConfig = globalVals.__SIMPLE_CONFIG__
 
-  test('class loads', () => {
+  it('class loads', () => {
     const siteScraper = new ScrapePages(simpleConfig)
     // siteScraper.run()
   })
